+++ changelog.md
@@ -19,6 +19,7 @@ This version is not yet released. If you are reading this on the website, then t
   - For example `⍜⇌°⊂` behaves in a more useful way
 - Allow [`on ⟜`](https://uiua.org/docs/on) to use function packs
 - Allow [`under ⍜`](https://uiua.org/docs/under) of scalar [`keep ▽`](https://uiua.org/docs/keep)
+- Add the [`graphemes`](https://uiua.org/docs/graphemes) function, which splits a string into unicode grapheme clusters
 - Add the [`&fmd`](https://uiua.org/docs/&fmd) system function, which creates a directory
 - Very large arrays are now displayed more concisely in output
 - Add unicode escape sequences for entering arbitrary characters